requirements.txt
google-genai
openai
transformers==4.51.3
ipdb